+++ crates/client/flashblocks-node/tests/state.rs
@@ -1,8 +1,6 @@
 //! Integration tests that stress Flashblocks state handling.
 
-use std::{sync::Arc, time::Duration};
-
-use alloy_consensus::{Receipt, Transaction};
+use alloy_consensus::Transaction;
 use alloy_eips::{BlockHashOrNumber, Encodable2718};
 use alloy_primitives::{Address, B256, BlockNumber, Bytes, U256, hex::FromHex, map::HashMap};
 use alloy_rpc_types_engine::PayloadId;
@@ -14,288 +12,11 @@ use base_flashblocks_node::test_harness::FlashblocksHarness;
 use base_flashtypes::{
     ExecutionPayloadBaseV1, ExecutionPayloadFlashblockDeltaV1, Flashblock, Metadata,
 };
-use op_alloy_consensus::OpDepositReceipt;
-use op_alloy_network::BlockResponse;
-use reth_chainspec::EthChainSpec;
-use reth_optimism_primitives::{OpBlock, OpReceipt, OpTransactionSigned};
-use reth_primitives_traits::{Account as RethAccount, Block as BlockT, RecoveredBlock};
-use reth_provider::{
-    AccountReader, BlockNumReader, BlockReader, ChainSpecProvider, StateProviderFactory,
-};
-use reth_transaction_pool::test_utils::TransactionBuilder;
-use tokio::time::sleep;
-// The amount of time to wait (in milliseconds) after sending a new flashblock or canonical block
-// so it can be processed by the state processor
-const SLEEP_TIME: u64 = 10;
-
-struct TestHarness {
-    node: FlashblocksHarness,
-    flashblocks: Arc<FlashblocksState>,
-    provider: LocalNodeProvider,
-}
-
-impl TestHarness {
-    async fn new() -> Self {
-        // These tests simulate pathological timing (missing receipts, reorgs, etc.), so we disable
-        // the automatic canonical listener and only apply blocks when the test explicitly requests it.
-        let node = FlashblocksHarness::manual_canonical()
-            .await
-            .expect("able to launch flashblocks harness");
-        let provider = node.blockchain_provider();
-        let flashblocks = node.flashblocks_state();
-
-        let genesis_block = provider
-            .block(BlockHashOrNumber::Number(0))
-            .expect("able to load block")
-            .expect("block exists")
-            .try_into_recovered()
-            .expect("able to recover block");
-        flashblocks.on_canonical_block_received(genesis_block);
-
-        Self { node, flashblocks, provider }
-    }
-
-    fn decode_private_key(account: Account) -> B256 {
-        B256::from_hex(account.private_key()).expect("valid hex-encoded key")
-    }
-
-    fn canonical_account(&self, account: Account) -> RethAccount {
-        self.provider
-            .basic_account(&account.address())
-            .expect("can lookup account state")
-            .expect("should be existing account state")
-    }
-
-    fn canonical_balance(&self, account: Account) -> U256 {
-        self.canonical_account(account).balance
-    }
-
-    fn expected_pending_balance(&self, account: Account, delta: u128) -> U256 {
-        self.canonical_balance(account) + U256::from(delta)
-    }
-
-    fn account_state(&self, account: Account) -> RethAccount {
-        let basic_account = self.canonical_account(account);
-
-        let nonce = self
-            .flashblocks
-            .get_pending_blocks()
-            .get_transaction_count(account.address())
-            .to::<u64>();
-        let balance = self
-            .flashblocks
-            .get_pending_blocks()
-            .get_balance(account.address())
-            .unwrap_or(basic_account.balance);
-
-        RethAccount {
-            nonce: nonce + basic_account.nonce,
-            balance,
-            bytecode_hash: basic_account.bytecode_hash,
-        }
-    }
-
-    fn build_transaction_to_send_eth(
-        &self,
-        from: Account,
-        to: Account,
-        amount: u128,
-    ) -> OpTransactionSigned {
-        let txn = TransactionBuilder::default()
-            .signer(Self::decode_private_key(from))
-            .chain_id(self.provider.chain_spec().chain_id())
-            .to(to.address())
-            .nonce(self.account_state(from).nonce)
-            .value(amount)
-            .gas_limit(21_000)
-            .max_fee_per_gas(1_000_000_000)
-            .max_priority_fee_per_gas(1_000_000_000)
-            .into_eip1559()
-            .as_eip1559()
-            .unwrap()
-            .clone();
-
-        OpTransactionSigned::Eip1559(txn)
-    }
-
-    fn build_transaction_to_send_eth_with_nonce(
-        &self,
-        from: Account,
-        to: Account,
-        amount: u128,
-        nonce: u64,
-    ) -> OpTransactionSigned {
-        let txn = TransactionBuilder::default()
-            .signer(Self::decode_private_key(from))
-            .chain_id(self.provider.chain_spec().chain_id())
-            .to(to.address())
-            .nonce(nonce)
-            .value(amount)
-            .gas_limit(21_000)
-            .max_fee_per_gas(1_000_000_000)
-            .max_priority_fee_per_gas(1_000_000_000)
-            .into_eip1559()
-            .as_eip1559()
-            .unwrap()
-            .clone();
-
-        OpTransactionSigned::Eip1559(txn)
-    }
-
-    async fn send_flashblock(&self, flashblock: Flashblock) {
-        self.node
-            .send_flashblock(flashblock)
-            .await
-            .expect("flashblocks channel should accept payload");
-        sleep(Duration::from_millis(SLEEP_TIME)).await;
-    }
-
-    async fn new_canonical_block_without_processing(
-        &mut self,
-        user_transactions: Vec<OpTransactionSigned>,
-    ) -> RecoveredBlock<OpBlock> {
-        let previous_tip =
-            self.provider.best_block_number().expect("able to read best block number");
-        let txs: Vec<Bytes> =
-            user_transactions.into_iter().map(|tx| tx.encoded_2718().into()).collect();
-        self.node.build_block_from_transactions(txs).await.expect("able to build block");
-        let target_block_number = previous_tip + 1;
-
-        let block = self
-            .provider
-            .block(BlockHashOrNumber::Number(target_block_number))
-            .expect("able to load block")
-            .expect("new canonical block should be available after building payload");
-
-        block.try_into_recovered().expect("able to recover newly built block")
-    }
-
-    async fn new_canonical_block(&mut self, user_transactions: Vec<OpTransactionSigned>) {
-        let block = self.new_canonical_block_without_processing(user_transactions).await;
-        self.flashblocks.on_canonical_block_received(block);
-        sleep(Duration::from_millis(SLEEP_TIME)).await;
-    }
-}
-
-struct FlashblockBuilder<'a> {
-    transactions: Vec<Bytes>,
-    receipts: Option<HashMap<B256, OpReceipt>>,
-    harness: &'a TestHarness,
-    canonical_block_number: Option<BlockNumber>,
-    index: u64,
-}
-
-impl<'a> FlashblockBuilder<'a> {
-    fn new_base(harness: &'a TestHarness) -> Self {
-        Self {
-            canonical_block_number: None,
-            transactions: vec![L1_BLOCK_INFO_DEPOSIT_TX],
-            receipts: Some({
-                let mut receipts = alloy_primitives::map::HashMap::default();
-                receipts.insert(
-                    L1_BLOCK_INFO_DEPOSIT_TX_HASH,
-                    OpReceipt::Deposit(OpDepositReceipt {
-                        inner: Receipt {
-                            status: true.into(),
-                            cumulative_gas_used: 10000,
-                            logs: vec![],
-                        },
-                        deposit_nonce: Some(4012991u64),
-                        deposit_receipt_version: None,
-                    }),
-                );
-                receipts
-            }),
-            index: 0,
-            harness,
-        }
-    }
-    fn new(harness: &'a TestHarness, index: u64) -> Self {
-        Self {
-            canonical_block_number: None,
-            transactions: Vec::new(),
-            receipts: Some(HashMap::default()),
-            harness,
-            index,
-        }
-    }
-
-    fn with_receipts(&mut self, receipts: Option<HashMap<B256, OpReceipt>>) -> &mut Self {
-        self.receipts = receipts;
-        self
-    }
-
-    fn with_transactions(&mut self, transactions: Vec<OpTransactionSigned>) -> &mut Self {
-        assert_ne!(self.index, 0, "Cannot set txns for initial flashblock");
-        self.transactions.clear();
-
-        let mut cumulative_gas_used = 0;
-        for txn in transactions.iter() {
-            cumulative_gas_used += txn.gas_limit();
-            self.transactions.push(txn.encoded_2718().into());
-            if let Some(ref mut receipts) = self.receipts {
-                receipts.insert(
-                    *txn.hash(),
-                    OpReceipt::Eip1559(Receipt {
-                        status: true.into(),
-                        cumulative_gas_used,
-                        logs: vec![],
-                    }),
-                );
-            }
-        }
-        self
-    }
-
-    const fn with_canonical_block_number(&mut self, num: BlockNumber) -> &mut Self {
-        self.canonical_block_number = Some(num);
-        self
-    }
-
-    fn build(&self) -> Flashblock {
-        let current_block = self.harness.node.latest_block();
-        let canonical_block_num =
-            self.canonical_block_number.unwrap_or_else(|| current_block.number) + 1;
-
-        let base = if self.index == 0 {
-            Some(ExecutionPayloadBaseV1 {
-                parent_beacon_block_root: current_block.hash(),
-                parent_hash: current_block.hash(),
-                fee_recipient: Address::random(),
-                prev_randao: B256::random(),
-                block_number: canonical_block_num,
-                gas_limit: current_block.gas_limit,
-                timestamp: current_block.timestamp + 2,
-                extra_data: Bytes::new(),
-                base_fee_per_gas: U256::from(100),
-            })
-        } else {
-            None
-        };
-
-        Flashblock {
-            payload_id: PayloadId::default(),
-            index: self.index,
-            base,
-            diff: ExecutionPayloadFlashblockDeltaV1 {
-                state_root: B256::default(),
-                receipts_root: B256::default(),
-                block_hash: B256::default(),
-                gas_used: 0,
-                withdrawals: Vec::new(),
-                logs_bloom: Default::default(),
-                withdrawals_root: Default::default(),
-                transactions: self.transactions.clone(),
-                blob_gas_used: Default::default(),
-            },
-            metadata: Metadata { block_number: canonical_block_num },
-        }
-    }
-}
+use base_flashblocks_node::test_harness::{FlashblockBuilder, FlashblocksBuilderTestHarness};
 
 #[tokio::test]
 async fn test_state_overrides_persisted_across_flashblocks() {
-    let test = TestHarness::new().await;
+    let test = FlashblocksBuilderTestHarness::new().await;
 
     test.send_flashblock(FlashblockBuilder::new_base(&test).build()).await;
     assert_eq!(
